requirements.txt
numpy
pyyaml
torch
gymnasium
mujoco
rich
